+++ docs/changelog.md
@@ -1,3 +1,5 @@
 ## v1.1
 - The game now reads from /dev/urandom to seed RNG, rather than using the system time.
 - The game now automatically uses your username for your save name. You can still use custom save names by passing '-s' when running the program.
+- You can now pass the '-d' argument with a value of 1-3 (inclusive) to bypass difficulty selection.
+- You can also pass '-h', '--help', or '--usage' to get a help document, or '-v' for version information.